requirements.txt
flask
pyyaml
wakeonlan